Finish unsynchronized message map
Still need to add more basic synchronization and then set up the
Readers-Writers design. Also need to add the code from
messagemap.go into main.go to test if the responses are working!
If so, then I'm almost done with the basic requirements for
submitting the task! Still want to do some unit testing and CI
to see how they work in go and GitHub.