Add TestPostRoot, need to change main methods
Made a unit test, but it fails because it relies on an uninitialized
messagemap. The messagemap normally is initialized in the `main()`
method of `main.go`, so it wouldn't usually be a problem. Now I'll
go and change the methods in `main.go` to take in a MessageMap as
a parameter (which is what I wanted to do in the beginning, I just
wasn't sure how).